ui/requirements.txt
Requests==2.32.4
authlib
psycopg2-binary==2.9.9
ruamel.yaml
gunicorn==21.2.0
python-dotenv>=1.0.0
Django>=4.2,<5
whitenoise>=6.6,<7
